Ensure expert_inputs are integer dtype for Embed layer

+++ VishwamAI/scripts/model_architecture.py
@@ -77,7 +77,7 @@ class VishwamAIModel(hk.Module):
             if jnp.any(mask):
                 print(f"mask shape: {mask.shape}, inputs shape: {inputs.shape}")  # Debugging print statement
                 mask = jnp.expand_dims(mask, axis=-1)  # Expand dimensions of mask to match inputs
-                expert_inputs = jnp.where(mask, inputs, 0).astype(jnp.float32)
+                expert_inputs = jnp.where(mask, inputs, 0)  # Ensure expert_inputs are integer dtype
                 expert_params = expert.init(jax.random.PRNGKey(42), expert_inputs)  # Initialize expert parameters
                 expert_outputs.append(expert.apply(expert_params, None, expert_inputs))  # Use apply method
